various warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@5095 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/HTTPServer/src/GuiElements.cpp
@@ -63,7 +63,7 @@ string sPageKeyword = szDefaultPageKeyword;
 
 void ReplaceAll(string &sSrc, const char * pszReplace, const string &sNew) {
 	string::size_type nCur = 0;
-	int nRepalceLen = strlen(pszReplace);
+	int nRepalceLen = (int)strlen(pszReplace);
 	while ((nCur = sSrc.find(pszReplace, nCur)) != sSrc.npos) {
 		sSrc.replace(nCur, nRepalceLen, sNew);
 		nCur += sNew.size();
